+++ test/runtests.jl
@@ -1,6 +1,9 @@
 module functions_tests
     include("functions.jl")
 end
+module chainrules_tests
+    include("chainrules.jl")
+end
 module explicit_imports_tests
     include("explicit_imports.jl")
 end